Fedrick

See merge request devland-id/batch-2/final-projects/group-2/wallet-api!2

+++ src/main/java/com/devland/walletapi/wallet/Wallet.java
@@ -5,6 +5,8 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
 import javax.persistence.*;
+
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -17,7 +19,7 @@ import java.util.List;
 public class Wallet {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private BigInteger id;
 
     private double walletBalance;
 
@@ -28,7 +30,7 @@ public class Wallet {
     @Column(nullable = false, updatable = false)
     private LocalDateTime walletCreatedAt;
 
-    public WalletResponseDTO convertToResponse(){
+    public WalletResponseDTO convertToResponse() {
         return WalletResponseDTO.builder().id(this.id)
                 .walletBalance(this.walletBalance)
                 .transactions(this.transactions)
