+++ src/main/java/com/devland/walletapi/wallet/WalletResponseDTO.java
@@ -6,16 +6,16 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
-
 
 @Getter
 @Setter
 @AllArgsConstructor
 @Builder
 public class WalletResponseDTO {
-    private Long id;
+    private BigInteger id;
 
     private double walletBalance;
